Don't clone constant ASTs when running with Analysis Driver.
This makes my Flutter benchmark another 4-5% faster.

Currently resolved ASTs and element models for them live and die
together. It is possible to misuse API, and keep element model extracted
from a resolved AST, so that this element model will continue keeping
the whole AST in memory. But there is also a way to get just element
model, without AST, so there is a way for the user to get it right.

Hm... It seems that either the element models we get are not complete
(and don't have ASTs for initializers, default values, etc). Or we don't
actually need to copy ASTs into element model, and the only reason
why we are doing this is that there is something in task model that
forced us to do this. We need to analyze why we do what we do while we
are removing task model.

R=brianwilkerson@google.com

Change-Id: I058c81cb99fe84c68e96e6bc722a27d87c499155
Reviewed-on: https://dart-review.googlesource.com/c/79641
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>