Eliminate an unnecessary boolean variable
There is a flag in the code switching between the current CFE/Analyzer
behavior and a different behavior.  It's not clear that this other
behavior is ever planned, but since we're not testing it it's better
to get rid of the code for now and restore it later if we ever want
it.

Change-Id: I8be4af7d39dcb9296b4b48fa10a6352df9e1e8b0
Reviewed-on: https://dart-review.googlesource.com/c/80660
Auto-Submit: Kevin Millikin <kmillikin@google.com>
Reviewed-by: Peter von der Ahé <ahe@google.com>
Commit-Queue: Peter von der Ahé <ahe@google.com>