Improve the way we call inferExpression
This function took a boolean flag typeNeeded and then it had an early
exit guarded by the condition (!typeNeeded && isTopLevel).  That is, a
false value for typeNeeded was just ignored when isTopLevel was false.

Change it so the early exit is only guarded by the value of the
argument, and ensure that the intended value is passed at all the call
sites.  This makes it easier to see under which conditions we are and
are not peforming type inference.

This is a prelimary step to simply not performing any inference when
we don't intend to perform any inference.

Change-Id: Ibe2260aa4967250afb8b23899216ae32c5321183
Reviewed-on: https://dart-review.googlesource.com/c/80667
Commit-Queue: Kevin Millikin <kmillikin@google.com>
Reviewed-by: Peter von der Ahé <ahe@google.com>