[vm] Fix race in assert in RawObject::VisitPointersPredefined.
While the concurrent marker is running RawObject::VisitPointersPredefined, the mutator may truncate the array being visited. This can cause the size in the object header and the size from the array length to temporarily disagree. Copy logic from RawObject::SizeFromClass to recompute the sizes in this case.

Bug: https://github.com/dart-lang/sdk/issues/22501
Bug: https://github.com/dart-lang/sdk/issues/34917
Change-Id: I2f1ffdc56ab69984a9d364cf0902f113d10f338b
Reviewed-on: https://dart-review.googlesource.com/c/81441
Commit-Queue: Ryan Macnak <rmacnak@google.com>
Reviewed-by: Vyacheslav Egorov <vegorov@google.com>