Set default type to Object when bound is explicitly Object.
This fixes the instantiated type for type parameters of function types
(whose default types are otherwise not computed) when the bound is
explicitly specified as Object.

It also enables error reporting to detect if a bound on a type variable
was specified or not, so that when a generic function type is printed,
the bound can be printed if and only if it was originally specified.

Fixes https://github.com/dart-lang/sdk/issues/34920

Change-Id: I9ad3f1cacb1b36eaef139d50c8e73cc7a8cc915b
Reviewed-on: https://dart-review.googlesource.com/c/81405
Reviewed-by: Dmitry Stefantsov <dmitryas@google.com>