[infra] Add approve_results.dart script for approving test result changes.
The approve_results program downloads the current results, the approved
results, and flakiness information for the specified bots. It presents
tables of the tests that have been fixed and the ones that have broken since
the last approval. Finally, it provides the option to approve the new
changes.

The approve_results program is part of the new workflow and only affects the
greenness of bots enrolled in it.

You can list the available bots by running:

    dart tools/approve_results.dart -l

You can approve the results on a bot by running:

    dart tools/approve_results.dart -b vm-kernel-linux-release-x64

Globs are supported, e.g. to approve all vm linux bots on ia32 and x64:

    dart tools/approve_results.dart -b 'vm-*linux*{ia32,x64}*'

By default the tool is interactive and asks whether the changes should be
approved. Approving the results turns the bots green on the next commit.
It's assumed that bugs have been filed prior to approving the test changes.
The -y option approves the changes non-interactively while the -n option
does a try run and shows what needs to be approved.

This tool is meant to be used early and often to turn the bot green. Emails
are not set about additional failures until the regressions has been
approved.

Bug: https://github.com/dart-lang/sdk/issues/35000
Change-Id: I6479edaa76556aa5ef499509c04b92ae770d1727
Reviewed-on: https://dart-review.googlesource.com/c/82222
Reviewed-by: William Hesse <whesse@google.com>