Fix viewing generated code in Observatory on Windows
For GCC and Clang, `long` and `unsigned long` are 64-bit on 64-bit platforms and 32-bit on 32-bit platforms, while on Windows they're always 32-bit. To fix this, I've changed `GetIntegerId` and `GetUnsignedIntegerId` to use the `long long` variants of `strtol` and `strtoul` on 64-bit platforms.

Closes #30201

Closes #35212
https://github.com/dart-lang/sdk/pull/35212

GitOrigin-RevId: 1b1ad64f9c02f24d1a81999e4ffb61951009abe8
Change-Id: Ieea5aaedc67d950dfb219981d2d8b878312d8315
Reviewed-on: https://dart-review.googlesource.com/c/84740
Reviewed-by: Ryan Macnak <rmacnak@google.com>
Commit-Queue: Ryan Macnak <rmacnak@google.com>