Cherry-pick analyzer CHANGELOG and pubspec from analyzer-0.33 branch.
Also includes changes to the kernel and front_end pubspecs, since
those packages are tightly coupled to the analyzer.

These changes were made on the analyzer-0.33 branch in preparation for
previous releases of the analyzer.  Merging them to the master branch
makes it easier to keep the pubspec and CHANGELOG files up to date on
that branch.

Change-Id: I08a7bdcd49b6b2396166953131927656eca2c563
Reviewed-on: https://dart-review.googlesource.com/c/85727
Auto-Submit: Paul Berry <paulberry@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>