[interpreter and dbc simulator] Remove usage counter increment of icdata owner.
Incrementing the usage counter of the calling function (icdata owner) during
an instance call does not seem the correct thing to do.
In kernel bytecode, the counter is incremented as part of executing the
CheckStack instruction occuring at the beginning of each function.
In Dart bytecode, the counter is incremented as part of executing the
HotCheck instruction.

Change-Id: I2e4eebc9ae682520b93518e76e1c6856d5d9bb78
Reviewed-on: https://dart-review.googlesource.com/c/86001
Reviewed-by: Alexander Markov <alexmarkov@google.com>
Commit-Queue: Régis Crelier <regis@google.com>