Cherry-pick be529425d24e6c7c0cbf6965456da1f2a3a96adf to analyzer-0.33 branch.
This change was intended for 0.33.x version of the analyzer.  It was
landed on master when master was on version 0.33.x and the analyzer
branch was on 0.34.x.  Now that the analyzer branch has been landed,
it needs to be cherry-picked to the analyzer-0.33 branch so that we
can use it to release an update to analyzer 0.33.x.

Original change description:
> Prepare for a future breaking change to the UriResolver class.
>
> See https://dart-review.googlesource.com/c/sdk/+/85726 for the
> breaking change that is anticipated.
>
> Change-Id: I4ba84a0e70bf5b3c62191c053ba9de7aee71ac93
> Reviewed-on: https://dart-review.googlesource.com/c/85728
> Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
> Commit-Queue: Paul Berry <paulberry@google.com>

Change-Id: I6db0d50dc7d342e071c226ae1046f7088ab4ec92
Reviewed-on: https://dart-review.googlesource.com/c/86503
Auto-Submit: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>