[vm/compiler] Do not use UnboxedIntConverter on tagged.
Rationale:
This avoids rewriting
   v1 <- Constant(#1)
   v2 <- BoxInt64:1402(v1)
   v3 <- UnboxInt64(v2)
into the non-existing
   v3 <- UnboxedIntConverter(tagged->int64, v1)

Note:
Arguably the input was incorrect (v1 should have been
an unboxed constant), but we allow type sloppy expressions
at a lot of places that are repaired later on, as for
this case after the fix. See the bug for more details.

https://github.com/dart-lang/sdk/issues/35371

Change-Id: I719124ba150a1c72fe6819a1edd8380847fd82c2
Reviewed-on: https://dart-review.googlesource.com/c/86823
Reviewed-by: Ryan Macnak <rmacnak@google.com>
Commit-Queue: Aart Bik <ajcbik@google.com>